+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerWithTimezone.test.jsx
@@ -5,11 +5,11 @@ const DateTimePickerExample = () => (
   <>
     <DateTimePickerTimezoneExampleTemplate
       value="2020-02-13T05:30:00"
-      timeZone="Africa/Bangui"
+      initialTimeZone="Africa/Bangui"
     />
     <DateTimePickerTimezoneExampleTemplate
       value="2017-11-05T01:30:00"
-      timeZone="Asia/Dubai"
+      initialTimeZone="Asia/Dubai"
     />
   </>
 );